salon/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.9.2
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
